Project 2 Reservoir data MongoDB
The reservoir data system supports storage of salinity and Calcium concentration (PPM)
values from various devices, located at strategic locations within a fresh water reservoir. It
also has a device registry.
The data is stored in MongoDB. The focus of the project is to learn some data modelling in
MongoDB and to implement various functionalities based on that.
Each device sends either salinity or Calcium PPM data based on its type. There is only one
category of user - admin.
To some degree, this project is constructed similar to the MongoDB database discussed in
the Jupyter notebook for the weekly content on NoSQL. You will find some hints and relevant
data modelling suggestions in that notebook.